Initial refactor to jpeg-based carousel
